feat(secrets): extract shared route validator + add github host
- single source of truth validateSecretRouteConfig in @stwd/vault
- reconcile drift between vault + api copies (stricter-wins)
- add api.github.com to host allowlist + proxy alias
- data-driven STRICT_HOSTS per-host narrowness guard
- merged-config revalidation on update (fail-closed)
