+++ packages/api/src/routes/secrets.ts
@@ -11,14 +11,13 @@
  * handlers to prevent Hono from treating "routes" as a secret ID.
  */
 
-import { isIP } from "node:net";
 import {
   getDb as getVaultDb,
   secretRoutes as secretRouteRows,
   secrets as secretRows,
   tenantConfigs as tenantConfigsTable,
 } from "@stwd/db";
-import { SecretVault } from "@stwd/vault";
+import { SecretVault, validateSecretRouteConfig } from "@stwd/vault";
 import { and, eq, inArray } from "drizzle-orm";
 import { type Context, Hono } from "hono";
 import { type AuditEventInput, writeAuditEvent } from "../services/audit";
@@ -52,29 +51,11 @@ async function writeSecretsAudit(
   });
 }
 
-const DEFAULT_SECRET_ROUTE_HOSTS = [
-  "api.openai.com",
-  "api.anthropic.com",
-  "public-api.birdeye.so",
-  "api.coingecko.com",
-  "api.helius.xyz",
-];
-const BLOCKED_INJECT_HEADERS = new Set([
-  "connection",
-  "content-length",
-  "host",
-  "proxy-authenticate",
-  "proxy-authorization",
-  "te",
-  "trailer",
-  "transfer-encoding",
-  "upgrade",
-]);
-const VALID_PROXY_METHODS = new Set(["*", "GET", "POST", "PUT", "PATCH", "DELETE", "HEAD"]);
-const ALLOW_BROAD_SECRET_ROUTES = process.env.STEWARD_ALLOW_BROAD_SECRET_ROUTES === "true";
-const MAX_SECRET_INJECT_FORMAT_LENGTH = 255;
+// Route-config validation (host allowlist, path/method/injectAs/injectKey/
+// injectFormat rules, per-host strictness) lives in the shared
+// validateSecretRouteConfig in @stwd/vault — the single source of truth also
+// used at the vault boundary. Only request-shape parsing stays local here.
 const MAX_SECRET_ROUTE_PRIORITY = 1_000_000;
-const HTTP_HEADER_NAME = /^[A-Za-z0-9!#$%&'*+\-.^_`|~]+$/;
 const SECRET_ROUTE_UPDATE_KEYS = new Set([
   "hostPattern",
   "agentId",
@@ -111,115 +92,6 @@ type SecretRouteCreate = {
   priority?: number;
   enabled?: boolean;
 };
-
-function configuredSecretRouteHosts(): string[] {
-  return [
-    ...DEFAULT_SECRET_ROUTE_HOSTS,
-    ...(process.env.STEWARD_SECRET_ROUTE_ALLOWED_HOSTS ?? "")
-      .split(",")
-      .map((host) => host.trim().toLowerCase())
-      .filter(Boolean),
-  ];
-}
-
-function hostAllowedByEntry(hostPattern: string, allowedHost: string): boolean {
-  if (hostPattern === allowedHost) return true;
-  if (hostPattern.startsWith("*.")) {
-    const suffix = hostPattern.slice(2);
-    const suffixLabels = suffix.split(".").filter(Boolean);
-    if (suffixLabels.length < 2) return false;
-    if (!allowedHost.startsWith("*.")) return false;
-    const allowedSuffix = allowedHost.slice(2);
-    return suffix === allowedSuffix || suffix.endsWith(`.${allowedSuffix}`);
-  }
-  if (allowedHost.startsWith("*.")) {
-    const allowedSuffix = allowedHost.slice(1);
-    return hostPattern.endsWith(allowedSuffix) && hostPattern.length > allowedSuffix.length;
-  }
-  return false;
-}
-
-function validateSecretRouteConfig(input: {
-  hostPattern?: string;
-  pathPattern?: string;
-  method?: string;
-  injectAs?: string;
-  injectKey?: string;
-  injectFormat?: string;
-}): string | null {
-  if (input.hostPattern !== undefined) {
-    const hostPattern = input.hostPattern.trim().toLowerCase();
-    if (!hostPattern || hostPattern === "*" || hostPattern === "*.*") {
-      return "hostPattern must be an explicit allowed host";
-    }
-    const hostForIpCheck = hostPattern.startsWith("*.") ? hostPattern.slice(2) : hostPattern;
-    if (
-      isIP(hostForIpCheck) ||
-      hostForIpCheck === "localhost" ||
-      hostForIpCheck.endsWith(".localhost") ||
-      hostForIpCheck.endsWith(".local") ||
-      hostForIpCheck.endsWith(".internal")
-    ) {
-      return "hostPattern must not target localhost, private, or internal hosts";
-    }
-    if (!configuredSecretRouteHosts().some((allowed) => hostAllowedByEntry(hostPattern, allowed))) {
-      return "hostPattern is not in the secret route allowlist";
-    }
-  }
-
-  if (input.pathPattern !== undefined) {
-    const pathPattern = input.pathPattern.trim();
-    const lowered = pathPattern.toLowerCase();
-    if (!pathPattern.startsWith("/")) return "pathPattern must start with /";
-    if (!ALLOW_BROAD_SECRET_ROUTES && (pathPattern === "/*" || pathPattern === "*")) {
-      return "broad pathPattern requires STEWARD_ALLOW_BROAD_SECRET_ROUTES=true";
-    }
-    if (/[\u0000-\u001f\u007f\\]/.test(pathPattern)) return "pathPattern is invalid";
-    if (
-      lowered.includes("%2e") ||
-      lowered.includes("%2f") ||
-      lowered.includes("%5c") ||
-      pathPattern.split("/").some((segment) => segment === "." || segment === "..")
-    ) {
-      return "pathPattern must not contain dot segments or encoded path separators";
-    }
-  }
-
-  if (input.method !== undefined) {
-    const method = input.method.trim().toUpperCase();
-    if (!VALID_PROXY_METHODS.has(method)) return "method is not allowed";
-    if (!ALLOW_BROAD_SECRET_ROUTES && method === "*") {
-      return "broad method requires STEWARD_ALLOW_BROAD_SECRET_ROUTES=true";
-    }
-  }
-
-  if (input.injectAs !== undefined) {
-    const validInjectAs = ["header"];
-    if (!validInjectAs.includes(input.injectAs)) {
-      return `'injectAs' must be one of: ${validInjectAs.join(", ")}`;
-    }
-  }
-
-  if (input.injectKey !== undefined) {
-    const key = input.injectKey.trim().toLowerCase();
-    if (!key || !HTTP_HEADER_NAME.test(key)) return "injectKey is invalid";
-    if (BLOCKED_INJECT_HEADERS.has(key)) return `injectKey '${input.injectKey}' is not allowed`;
-  }
-
-  if (input.injectFormat !== undefined) {
-    if (typeof input.injectFormat !== "string") return "injectFormat must be a string";
-    if (input.injectFormat.length > MAX_SECRET_INJECT_FORMAT_LENGTH) {
-      return `injectFormat cannot exceed ${MAX_SECRET_INJECT_FORMAT_LENGTH} characters`;
-    }
-    if (/[\r\n]/.test(input.injectFormat)) return "injectFormat must not contain line breaks";
-    const placeholderCount = input.injectFormat.match(/\{value\}/g)?.length ?? 0;
-    if (placeholderCount !== 1) {
-      return "injectFormat must contain exactly one {value} placeholder";
-    }
-  }
-
-  return null;
-}
 
 function parseSecretRouteUpdate(body: Record<string, unknown>):
   | {
@@ -655,6 +527,24 @@ secretsRoutes.put("/routes/:id", async (c) => {
   const existing = await sv.getRoute(tenantId, routeId);
   if (!existing) {
     return c.json<ApiResponse>({ ok: false, error: "Route not found" }, 404);
+  }
+  // Fail-closed re-validation against the MERGED config. A partial update (e.g.
+  // changing only pathPattern or method) is validated in isolation above and so
+  // would not trigger per-host strictness for a route that already targets a
+  // strict host. Re-run the shared validator on existing ∪ update so a strict
+  // host's narrowness (explicit method + deep path) can never be loosened by a
+  // partial edit.
+  const mergedConfig = {
+    hostPattern: update.hostPattern ?? existing.hostPattern ?? undefined,
+    pathPattern: update.pathPattern ?? existing.pathPattern ?? undefined,
+    method: update.method ?? existing.method ?? undefined,
+    injectAs: update.injectAs ?? existing.injectAs ?? undefined,
+    injectKey: update.injectKey ?? existing.injectKey ?? undefined,
+    injectFormat: update.injectFormat ?? existing.injectFormat ?? undefined,
+  };
+  const mergedValidationError = validateSecretRouteConfig(mergedConfig);
+  if (mergedValidationError) {
+    return c.json<ApiResponse>({ ok: false, error: mergedValidationError }, 400);
   }
   await writeSecretsAudit(c, {
     tenantId,
