+++ packages/proxy/src/config.ts
@@ -13,6 +13,7 @@ export const DEFAULT_ALIASES: Record<string, string> = {
   birdeye: "public-api.birdeye.so",
   coingecko: "api.coingecko.com",
   helius: "api.helius.xyz",
+  github: "api.github.com",
 };
 
 /** Default port for the proxy server */
